+++ CHANGELOG.md
@@ -5,6 +5,11 @@ The most recent release is at the top.
 
 ## `0.2.0-snapshot`
 
+### Asset export
+
+- Add new flag `-H`/`--include-hidden` to include hidden assets in the export.  
+By default, hidden assets are not included in the export.
+
 ## `0.1.0`
 
 Initial release.